+++ lib/facebook_ads/ad_objects/ad_study.rb
@@ -30,6 +30,7 @@ module FacebookAds
       "CONTINUOUS_LIFT_CONFIG",
       "GEO_LIFT",
       "LIFT",
+      "PRIVATE_LIFT",
       "SPLIT_TEST",
     ]