+++ lib/facebook_ads/ad_objects/campaign.rb
@@ -69,7 +69,6 @@ module FacebookAds
       "last_week_sun_sat",
       "last_year",
       "lifetime",
-      "maximum",
       "this_month",
       "this_quarter",
       "this_week_mon_today",