+++ lib/facebook_ads/ad_objects/da_check.rb
@@ -26,6 +26,13 @@ module FacebookAds
   # pull request for this class.
 
   class DaCheck < AdObject
+    CONNECTION_METHOD = [
+      "ALL",
+      "APP",
+      "BROWSER",
+      "SERVER",
+    ]
+
 
     field :action_uri, 'string'
     field :description, 'string'