+++ lib/facebook_ads/ad_objects/photo.rb
@@ -125,6 +125,12 @@ module FacebookAds
       end
     end
 
+    has_edge :reactions do |edge|
+      edge.get 'Profile' do |api|
+        api.has_param :type, { enum: -> { Profile::TYPE }}
+      end
+    end
+
     has_edge :sharedposts do |edge|
       edge.get 'Post'
     end